Refactored layout of QQS and transition animation.
QuickQSPanel.HeaderTileLayout is reworked to work in displays of
different densities by inheriting from TileLayout. The maximum number of
tiles shown is still determined by the same parameter, but it can show
less tiles if there is not enough space due to low density/small screen.
This is calculated during onMeasure using the available space.

QSAnimator is modified to accomodate the following three cases:

* If the tile is present in QQS and QS, animate as before.
* If the tile is present in QQS and not in QS, fade translate to right.
* If the tile is not present in QQS and is in QS, fade in from the
right.

Bug: 80141553
Bug: 110404680
Change-Id: I7be192b9ad6e66f97b3041bd8d5ac57ad71e6fdd
Test: Visual, different densities and orientations.