Change the feature flag for Hearing Aid behavior
Previously, this feature flag for Hearing Aid Profile is to disable the
profile on supported phone . This CL will reverse the logic and is
to enable the Hearing Aid profile on unsupported phone. This is to allow
testing of this profile on unsupported phone.

Test: Manual tests on different types of phones
Bug: 116861680
Change-Id: Ic8db4445a58942c5f75598dce0503f6670ff9c7b
Merged-In: Ic8db4445a58942c5f75598dce0503f6670ff9c7b
(cherry picked from commit 75d4929b5d72235bbc57673d2681919569a39483)