Tell KeygaurdViewMediator about keyguard showing updates
StatusBarKeygaurdViewManager used to rely on StatusBar to tell
KeyguardMonitor about keyguard show/hide, so this change makes the
manager tell the monitor directly about updates.

The problem with the old way is that the keyguard show/hide state comes
down from PhoneWindowManager, and StatusBar was only updating the state
on StatusBarState change. StatusBarState is a sysui-only state, and
doesn't adequately track window manager state.

Fixes: 116806599
Test: unlock phone with fingerprint and swipe for notifications during
the animation.

Change-Id: I220643e115880ff29df92e1fcc76dfe0424a00f6