Add ModuleInfo class
The ModuleInfo class contains details about mainline modules. We're
not sure about what kind of information we'll need, so, creating
a placeholder structure and accessor methods that can be adjusted
later.

While the package manager does not know anything about what a "module"
is, we use the package manager as convenience to not creating a new
MainlineManager class.

Bug: 119220828
Test: Workspace still builds
Change-Id: I42cbc5e119652edd3eda155ddef861e3e0889479