Introduce android:preferCodeIntegrity, private for now
This attribute is default to false, and if set to true, indicates that
the app wants to run with strong integrity guarantee currently the
platform can provide the best.

In this change, this flag implies that on install time,
 1) .dex and .so must be stored uncompressed and aligned to install
 2) android:extractNativeLibs must be false to install

At run time, ART will run from the dex within the APK directly, and NDK
libraries will be mapped directly from the APK.  This way, thest files
stay protected by signature.

The attribute currently stays in private to make development progress.
We plan to make it public once we confirm the demand from some apps.

Test: atest AppIntegrityTest (to be added in ag/5554864 after publicized)
Bug: 112037137
Change-Id: Ifde90cb0666fbb57e8b61f90b4ba1a2dd2a2b4ae