Let shared user checking to be backward-compatible
The shared user checking becomes strict. If there is an APK in
vendor partition shared UID with a APK in system partition, the
signatures of these 2 APKs must be equal. Otherwise, framework
will throw an exception and cannot boot.

After Android 9, all devices must make sure their APKs passing
the checking.

The patch relaxes the exception for the devices which upgraded
from 9 or earlier version. With this patch, we still can boot the
upgraded devices with old vendor and new GSI. For example,
vendor on 9 and GSI on 10.

Bug: 117135529
Test: boot a walleye device on P with aosp_arm64_ab-userdebug
      (SPL patched) with the patch

Change-Id: I55287210e226ad760b761eb624a1826dc7e31bd8