[AWARE] Add builder pattern for Wi-Fi Aware NetworkSpecifier
The NetworkSpecifier is required when requesting a Wi-Fi Aware
data-path link. Previously the NetworkSpecifier was created using
3 separate methods - such a mechanism is not scalable as we need
to add more parameters (i.e. more methods).

Add Builder pattern for creating the Aware NetworkSpecifier and
deprecate the original methods.

This CL does not add any new parameter - it should be functionally
the same as the original codebase.

Bug: 117605977
Test: atest android.net.wifi.aware
Test: ACTS ThroughputTest:test_iperf_single_ndp_aware_only_ib
Change-Id: Ifd4a1791a5fa0c351912733012e543bd90939c7f